Create high performance sync integration for Elastic Search
The code has been extracted from an existing integration with a different
search provider and adapted to work with Elastic Search.

The extension allows for high performance serialization and indexing of
products through concurrent background jobs.

To achieve that, serialized data is kept in a separate table wrapped by
the `Solidus::Elastic::Product::State model`.

Integration with Elastic Model library
--------------------------------------

This extension implements the full interface of the official
[Elastic Model library](https://github.com/elastic/elasticsearch-rails/tree/master/elasticsearch-model)
through the `State` and `Index` classes.Access to Elastic

  - Use the `Index` class to access Elastic's class methods, i.e. `client`,
    `search`, `import`, `document_type`, `index_name`

  - Use the `State` class to access Elastic's instance methods, i.e.
    `index_document`, `update_document`, `delete_document`